refactor: created new productsContainer component to swicth between the cusotmer and admin conatiner
